Refactor attachment processing and introduce Azure Blob storage support
